+++ include/chronon3d/backends/software/software_renderer.hpp
@@ -148,6 +148,7 @@ public:
     Camera2_5D m_prev_camera;
     bool m_prev_camera_valid{false};
     uint64_t m_prev_scene_fingerprint{0};
+    uint64_t m_prev_static_scene_fingerprint{0};
     graph::SceneHasher m_scene_hasher;
 
     [[nodiscard]] int last_layer_count() const { return m_last_layer_count; }