+++ include/chronon3d/core/profiling/counters.hpp
@@ -58,6 +58,7 @@ constexpr std::size_t kHardwareDestructiveInterferenceSize = 64;
     X(dirty_pixels) \
     X(dirty_union_area_pixels) \
     X(dirty_full_fallbacks) \
+    X(execution_plan_cache_hits) \
     X(bypass_not_cacheable_count) \
     X(framebuffer_acquire_ms) \
     X(framebuffer_clear_ms) \